Address PR review: fix registry sync, weight validation, and per-chunk allocation
- Dedup herbarium entries by entry.name (not species.name) so user-created
  plants are distinct; set species.name in new_entry to match entry name
- Rebuild renderer registry (prototype meshes) at game start so edits to
  the herbarium are reflected in both runtime and renderer
- Build PlantRegistry once and share via Arc instead of rebuilding per-chunk;
  propagate Arc<PlantRegistry> through ChunkGenerator/ContentLayer/FloraLayer
- Clamp placement weights to non-negative and skip when total_weight <= 0
- Reuse scratch buffer for eligible-species filtering instead of allocating
  a new Vec per grid cell
- Remove unnecessary usize cast of species_index

+++ src/renderer_wgpu/instancing.rs
@@ -106,7 +106,7 @@ pub fn build_plant_instances(
     let mut per_species: Vec<Vec<InstanceData>> = vec![Vec::new(); species.len()];
 
     for p in plants {
-        let idx = p.species_index as usize;
+        let idx = p.species_index;
         if idx >= species.len() {
             continue;
         }
